refactored process
