+++ src/index.tsx
@@ -5,6 +5,7 @@ import React from "react";
 import * as ReactDOMClient from "react-dom/client";
 import { Provider } from "react-redux";
 
+
 import configureStore from "@_redux/_root";
 
 import App from "./app";
